Rename table package to uncomfortable
