+++ src/main/java/com/moment/the/answer/service/AnswerService.java
@@ -8,8 +8,8 @@ import com.moment.the.answer.dto.AnswerDto;
 import com.moment.the.answer.dto.AnswerResDto;
 import com.moment.the.answer.repository.AnswerRepository;
 import com.moment.the.exceptionAdvice.exception.*;
-import com.moment.the.table.TableDomain;
-import com.moment.the.table.repository.TableRepository;
+import com.moment.the.uncomfortable.TableDomain;
+import com.moment.the.uncomfortable.repository.TableRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
