+++ src/test/java/com/moment/the/controller/release/TableControllerTest.java
@@ -2,13 +2,13 @@ package com.moment.the.controller.release;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.moment.the.table.controller.TableController;
-import com.moment.the.table.TableDomain;
+import com.moment.the.uncomfortable.controller.TableController;
+import com.moment.the.uncomfortable.TableDomain;
 import com.moment.the.response.ResponseService;
-import com.moment.the.table.dto.TableDto;
-import com.moment.the.table.dto.TableViewDto;
-import com.moment.the.table.repository.TableRepository;
-import com.moment.the.table.service.TableService;
+import com.moment.the.uncomfortable.dto.TableDto;
+import com.moment.the.uncomfortable.dto.TableViewDto;
+import com.moment.the.uncomfortable.repository.TableRepository;
+import com.moment.the.uncomfortable.service.TableService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.*;
