+++ src/test/java/com/moment/the/service/AnswerServiceTest.java
@@ -9,11 +9,11 @@ import com.moment.the.answer.repository.AnswerRepository;
 import com.moment.the.answer.service.AnswerService;
 import com.moment.the.config.security.auth.MyUserDetailsService;
 import com.moment.the.admin.AdminDomain;
-import com.moment.the.table.TableDomain;
+import com.moment.the.uncomfortable.TableDomain;
 import com.moment.the.admin.dto.AdminDto;
-import com.moment.the.table.dto.TableDto;
+import com.moment.the.uncomfortable.dto.TableDto;
 import com.moment.the.admin.repository.AdminRepository;
-import com.moment.the.table.service.TableService;
+import com.moment.the.uncomfortable.service.TableService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
